requirements.txt
Flask
requests
sentence-transformers
pinecone-client
